+++ src/frequenz/sdk/timeseries/formula_engine/_formula_generators/_grid_reactive_power_formula.py
@@ -1,47 +1,47 @@
 # License: MIT
-# Copyright © 2022 Frequenz Energy-as-a-Service GmbH
+# Copyright © 2024 Frequenz Energy-as-a-Service GmbH
 
-"""Formula generator from component graph for Grid Power."""
+"""Formula generator from component graph for Grid Reactive Power."""
 
 
 from frequenz.client.microgrid import Component, ComponentMetricId
-from frequenz.quantities import Power
+from frequenz.quantities import ReactivePower
 
 from .._formula_engine import FormulaEngine
 from ._fallback_formula_metric_fetcher import FallbackFormulaMetricFetcher
 from ._formula_generator import FormulaGeneratorConfig
 from ._grid_power_formula_base import GridPowerFormulaBase
-from ._simple_formula import SimplePowerFormula
+from ._simple_formula import SimpleReactivePowerFormula
 
 
-class GridPowerFormula(GridPowerFormulaBase[Power]):
-    """Creates a formula engine from the component graph for calculating grid power."""
+class GridReactivePowerFormula(GridPowerFormulaBase[ReactivePower]):
+    """Creates a formula engine from the component graph for calculating grid reactive power."""
 
     def generate(  # noqa: DOC502
         # * ComponentNotFound is raised indirectly by _get_grid_component_successors
         self,
-    ) -> FormulaEngine[Power]:
-        """Generate a formula for calculating grid power from the component graph.
+    ) -> FormulaEngine[ReactivePower]:
+        """Generate a formula for calculating grid reactive power from the component graph.
 
         Returns:
-            A formula engine that will calculate grid power values.
+            A formula engine that will calculate grid reactive power values.
 
         Raises:
             ComponentNotFound: when the component graph doesn't have a `GRID` component.
         """
         builder = self._get_builder(
-            "grid-power",
-            ComponentMetricId.ACTIVE_POWER,
-            Power.from_watts,
+            "grid_reactive_power_formula",
+            ComponentMetricId.REACTIVE_POWER,
+            ReactivePower.from_volt_amperes_reactive,
         )
         return self._generate(builder)
 
     def _get_fallback_formulas(
         self, components: set[Component]
-    ) -> dict[Component, FallbackFormulaMetricFetcher[Power] | None]:
+    ) -> dict[Component, FallbackFormulaMetricFetcher[ReactivePower] | None]:
         """Find primary and fallback components and create fallback formulas.
 
-        The primary component is the one that will be used to calculate the grid power.
+        The primary component is the one that will be used to calculate the grid reactive power.
         If it is not available, the fallback formula will be used instead.
         Fallback formulas calculate the grid power using the fallback components.
         Fallback formulas are wrapped in `FallbackFormulaMetricFetcher`.
@@ -55,7 +55,7 @@ class GridPowerFormula(GridPowerFormulaBase[Power]):
         fallbacks = self._get_metric_fallback_components(components)
 
         fallback_formulas: dict[
-            Component, FallbackFormulaMetricFetcher[Power] | None
+            Component, FallbackFormulaMetricFetcher[ReactivePower] | None
         ] = {}
 
         for primary_component, fallback_components in fallbacks.items():
@@ -64,7 +64,7 @@ class GridPowerFormula(GridPowerFormulaBase[Power]):
                 continue
 
             fallback_ids = [c.component_id for c in fallback_components]
-            generator = SimplePowerFormula(
+            generator = SimpleReactivePowerFormula(
                 f"{self._namespace}_fallback_{fallback_ids}",
                 self._channel_registry,
                 self._resampler_subscription_sender,