+++ src/frequenz/sdk/timeseries/formula_engine/_formula_generators/_producer_power_formula.py
@@ -17,7 +17,7 @@ from ._formula_generator import (
     FormulaGenerator,
     FormulaGeneratorConfig,
 )
-from ._simple_power_formula import SimplePowerFormula
+from ._simple_formula import SimplePowerFormula
 
 _logger = logging.getLogger(__name__)